Reset preferred codec options on stream disconnect
